refactor: 직관 내역 관련 model 패키지 정리
