+++ android/app/src/main/java/com/yagubogu/presentation/attendance/AttendanceHistoryViewHolder.kt
@@ -4,12 +4,13 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.yagubogu.databinding.ItemAttendanceHistorySummaryBinding
+import com.yagubogu.presentation.attendance.model.AttendanceHistoryItem
 
 class AttendanceHistoryViewHolder(
     private val binding: ItemAttendanceHistorySummaryBinding,
 ) : RecyclerView.ViewHolder(binding.root) {
     fun bind(item: AttendanceHistoryItem) {
-        binding.attendanceHistoryItem = item
+        binding.layoutAttendanceHistory.attendanceHistoryItem = item
     }
 
     companion object {
