+++ android/app/src/main/java/com/yagubogu/presentation/attendance/model/AttendanceHistoryFilter.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.presentation.attendance
+package com.yagubogu.presentation.attendance.model
 
 enum class AttendanceHistoryFilter {
     ALL,
