+++ android/app/src/main/java/com/yagubogu/presentation/attendance/model/AttendanceHistoryTeamItem.kt
@@ -1,10 +1,10 @@
-package com.yagubogu.presentation.attendance
+package com.yagubogu.presentation.attendance.model
 
 import androidx.annotation.ColorRes
 import com.yagubogu.domain.model.Team
 import com.yagubogu.presentation.util.getTeamColor
 
-data class CheckInGameTeamItem(
+data class AttendanceHistoryTeamItem(
     val team: Team,
     val name: String,
     val score: Int,
