+++ android/app/src/main/res/layout/layout_attendance_history.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="attendanceHistoryItem"
-            type="com.yagubogu.presentation.attendance.AttendanceHistoryItem" />
+            type="com.yagubogu.presentation.attendance.model.AttendanceHistoryItem" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
